Add files via upload
Problem#2 - Quick Sort
Given an array A of integers and an integer k, you're to get the kth minimum integer in
this array.
Since the array size maybe large, the array is not given as input. You've to generate it
yourself. To generate the array elements, consider this function that generates a random
number: