Add files via upload
You and your friends are building a smart social networking system (choose its name). You
are going to compete with the facebook app. Competition is very hard, and you need to
build strong features. One of the most critical parts of this system is user manager, and they
decided you are the one to develop it.